feat: lost bids, date time in auctionC closing summary

+++ src/main/java/com/weareadaptive/auctionhouse/model/Bid.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 import static com.weareadaptive.auctionhouse.StringUtil.isNullOrEmpty;
 
-public record Bid(String owner, BigDecimal price, int quantity, long submissionTime) {
+public record Bid(String owner, double price, int quantity, long submissionTime) {
     public Bid {
         if (isNullOrEmpty(owner)) {
             throw new BusinessException("owner cannot be null or empty");
